+++ src/components/TaskList.js
@@ -1,0 +1,9 @@
+export default function TaskList({tasks}) {
+  return (
+    <ul>
+      {tasks.map(task => (
+        <li key={task}>{task}</li>
+      ))}
+    </ul>
+  );
+}
